api modifications

+++ Rpjsip/pjmanager.h
@@ -2,6 +2,6 @@ int init(char *domain, char *user, char *passwd, char *proxy);
 
 void destroy();
 
-int call( long acc_id, char* number, char* domain );
+int call( int acc_id, char* number, char* domain );
 
 void endCall(int call_id);
